chore: remove redundant shared signals int tests
